canny-ml-service/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.0
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.4
